modifié :         inc/fonctions.h 	modifié :         src/fonctions.c 	modifié :         src/main.c

+++ inc/fonctions.h
@@ -26,10 +26,10 @@
 #define RAYON 1
 
 int k,direction;
-int xs,ys,zs;
+double xs,ys,zs;
 double angle;
 double increment;
-int snake[TAILLE_MAX][3];
+double snake[TAILLE_MAX][3];
 double vitesse;
 int compteur;
 int xSouris,ySouris;
